lint

+++ packages/axelar-core-std/src/nexus/execute.rs
@@ -102,7 +102,7 @@ mod test {
             destination_address: "something else".parse().unwrap(),
             payload_hash: [1; 32],
             source_tx_id: msg_id.tx_hash.to_vec().try_into().unwrap(),
-            source_tx_index: msg_id.event_index as u64,
+            source_tx_index: msg_id.event_index,
             id: msg_id.to_string(),
         };
